Rename "generator" mod to "markdown_generator"
